RateLimiting tests: fold typed/guard/external-limiter into existing file
